🎨 [refactor] move autoloader to dedicate folder
